+++ src/components/pages/campaigns-create-new/styled-components.tsx
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import { Button, Radio, Input, Switcher } from 'components/common'
+import { Button, Radio, Input, Switcher, Select } from 'components/common'
 
 export const WidgetButton = styled(Button)`
   padding: 4px 50px;
@@ -12,5 +12,9 @@ export const StyledRadio = styled(Radio)`
 export const InputStyled = styled(Input)``
 
 export const SwitcherStyled = styled(Switcher)`
+  
+`
+
+export const SelectStyled = styled(Select)`
   
 `