+++ src/helpers/index.js
@@ -57,9 +57,15 @@ import getContractVersion from './get-contract-version'
 import createEncryptionKey from './create-encryption-key'
 import countNativeTokensToSecure from './count-native-tokens-to-secure'
 import defineSystem from './define-system'
+import convertAlchemyTokens from './convert-alchemy-tokens'
+import defineTokenType from './define-token-type'
+import defineAlchemyNetwork from './define-alchemy-network'
 
 export {
   shortenString,
+  defineTokenType,
+  defineAlchemyNetwork,
+  convertAlchemyTokens,
   defineSystem,
   countNativeTokensToSecure,
   getContractVersion,